Add a script to build libraries with Android NDK.
This script automates the process of building open source shared libraries
with the Android Native Development Kit (NDK). The script is meant to be
included from per-library build scripts.

The script attempts to detect as much as possible about the NDK. It can also
download the source tarball for the library if a URL was provided by the
caller. After the library has been built, the script updates it's SONAME,
removing the version component from the end. This hack is, unfortunately,
necessary to overcome limitations in the Android linker and the APK builder.