Add a build script and patches for libmicrohttpd.
The build script makes it possible to build libmicrohttpd on Android. Since
the library would not build cleanly, we also add a few patches that make it
possible to build vanilla sources of libmicrohttpd with Android NDK. Those
patches will be applied with quilt after the build script has extracted the
source tarball.