Also install libraries into a common sysroot.
To make it possible to link against libraries cross-compiled for Android,
we also install all cross-compiled libraries into a common sysroot directory,
in addition to a per-library directory as before. That way all pkg-config
and also all libraries will be in the same directory, just like on a real
system.

To make this possible, the function fix_soname needs to be updated to fix
SONAMES in libraries in place. The function replaces the .so symbolic link
with a copy of the library with updated SONAME.