Add Android build script for libev.
Libev is configured to provide its own version of floor and skip
linking with -lm. Although this is technically not necessary, it
simplifies building and linking the final application.

Also, fd_mask does not appear to be available on Android so we
define it to int. That's a hack, but it'll do for now.